added gooey GUI elements
