+++ FlexTools/Collections/Examples.ini
@@ -1,9 +1,6 @@
-[Reports.Text Statistics]
+[OSM.OSM]
 _order = 1
 
-[Reports.Lexicon Statistics]
+[SetString.SetString]
 _order = 2
-
-[OSM.OSM]
-_order = 3
