+++ FlexTools/flextools.ini
@@ -1,4 +1,4 @@
 [DEFAULT]
 currentcollection = u'Examples'
-currentdatabase = u'RuthTestFLExTools'
+currentdatabase = u'testfindreplace2'
